Changed redirect header from log_off.php to ../login.php
previous step had to redirected to log_off.php page again after the
session has been distroyed with a notice that you have been logged off
completely. This modification is to redirect back to the public
login.php and then alerts a message that the user was successfully
logout and that they can use the form below to login again.