Added Find space and List space to the user account section
The next step is to create the user favourites and their submitted
listing. This will be achieved by adding one extra tables called
favourites so that users can delete and mark their favourites. For each
listing we will need to have a user who submitted it. This user can be
the admin, public or user with an account ID. For the public we can
record session information. and that will be stored agains each office
listing record. Will need to decide later if users can submit building
or should only submit listing made by admin - This case can be taken
care of by allowing users to add address and later create a listing in
the step process. This will be done hence when a user search for an
addresss they will be responded that the address is not available and a
question will be asked of them if they wish to add the address. If they
select yes if its an address that has been validated by google then the
address will be addredd with the building information to the database.
Enough of logic, commit and back to coding